handle queue removal

+++ sandbox/sandbox_queue.py
@@ -5,6 +5,7 @@ mines crypto whenever queue is empty
 import subprocess
 import multiprocessing
 from flask import Flask, jsonify, request
+from flask_apscheduler import APScheduler
 import os
 
 
@@ -54,23 +55,46 @@ def push_job(params):
     with open(f"{job_dir}/render_file.blend", "w") as f:
         f.write(render_file)
     
+    # append job to queue first to prevent mining from starting after the stop call
+    job = {"job_dir": job_dir, "job_id": job_id, "tsp_id": None}
+    global QUEUE
+    QUEUE.append(job)
+    # make sure mining is stopped before running render job
     stop_mining()
     tsp_id = run_shell_cmd(f"tsp python3 run.py").strip()
-    global QUEUE
-    QUEUE.append({"job_dir": job_dir, "tsp_id": tsp_id})
-    
+    # since this is a reference to the job in QUEUE, this sets tsp_id in QUEUE
+    job["tsp_id"] = tsp_id
+
 
 def pop_job(params):
     """
     remove job from queue
+    does nothing if already removed from queue
     """
-    # TODO figure out id based on params
-    tsp_id = ""
+    job_id = params["job_id"]
+    queued_job = None
+    for i, job in enumerate(QUEUE):
+        if job["job_id"] == job_id:
+            queued_job = i
+            break
+    # job already removed
+    if queued_job is None:
+        return
+    # remove relevant job from queue
+    global QUEUE
+    queued_job = QUEUE.pop(queued_job)
+    tsp_id = queued_job["tsp_id"]
     pid = run_shell_cmd(f"tsp -p {tsp_id}").strip()
     run_shell_cmd(f"kill -9 {pid}")
-    # TODO stop tracking tsp_id
 
 
+def handle_finished_jobs():
+    """
+    checks for any finished jobs and sends results back to host daemon
+    cleans up and removes files afterwards
+    starts crypto miner if all jobs are finished
+    """
+    
 def run_flask_server(q):
     @app.route("/", methods=["POST"])
     def index():
@@ -100,6 +124,10 @@ FILE_DIR = "~/jobs"
 def main():
     start_mining()
     app.secret_key = uuid.uuid4().hex
+    # create a scheduler that periodically checks/handles finished jobs starts mining when there are no jobs in queue
+    scheduler = APScheduler()
+    scheduler.add_job(id='Start Mining', func=handle_finished_jobs, trigger="interval", seconds=5)
+    scheduler.start()
     q = multiprocessing.Queue()
     server = multiprocessing.Process(target=run_flask_server, args=(q,))
     server.start()
